# docs/api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "QuickFlow API",
    "description": "Tasks, habits, learning resources, todo plans, settings and dashboard for the QuickFlow personal productivity app.",
    "version": "v1"
  },
  "paths": {
...
    "/api/plans": {
      "get": {
        "tags": [
          "Plans"
        ],
        "summary": "Lists plans ordered by priority order (then start time), optionally filtered by computed status.",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/PlanStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Plans"
        ],
        "summary": "Creates a plan from existing tasks, habits and/or learning resources.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/plans/{id}/notifications/ack": {
      "post": {
        "tags": [
          "Plans"
        ],
        "summary": "Acknowledges a plan's start notification so it is not reported again.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/{id}": {
      "get": {
        "tags": [
          "Plans"
        ],
        "summary": "Gets one plan with its items and computed progress.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Plans"
        ],
        "summary": "Updates a plan's fields and items (retained items keep their done flag).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Plans"
        ],
        "summary": "Removes a plan (its items are removed; the referenced tasks/habits/cards are kept).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/{id}/items/{itemId}": {
      "patch": {
        "tags": [
          "Plans"
        ],
        "summary": "Marks a plan item done or not done. Returns the updated plan (progress recomputed).\r\nThe source entity's own completion follows (BR-13): task Done/Todo, habit completion for today, card Completed/InProgress.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanItemRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanItemRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PlanItemRequest": {
        "type": "object",
        "properties": {
          "sourceType": {
            "$ref": "#/components/schemas/PlanItemSourceType"
          },
          "sourceId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Reference to an existing entity to include in a plan."
      },
      "PlanItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "planId": {
            "type": "integer",
            "format": "int32"
          },
          "sourceType": {
            "$ref": "#/components/schemas/PlanItemSourceType"
          },
          "sourceId": {
            "type": "integer",
            "format": "int32"
          },
          "sourceTitle": {
            "type": "string",
            "nullable": true
          },
          "isDone": {
            "type": "boolean"
          }
        },
        "additionalProperties": false,
        "description": "An item of a plan. `SourceTitle` is the current title/name of the referenced entity."
      },
      "PlanItemSourceType": {
        "enum": [
          "Task",
          "Habit",
          "LearningResource"
        ],
        "type": "string"
      },
      "PlanRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Required, at most 200 characters.",
            "nullable": true
          },
          "estimatedDurationMinutes": {
            "type": "integer",
            "description": "Estimated duration in minutes; required, greater than 0.",
            "format": "int32",
            "nullable": true
          },
          "startDateTime": {
            "type": "string",
            "description": "Start date-time (ISO 8601). Values without an offset are taken as server local time.",
            "format": "date-time",
            "nullable": true
          },
          "endDateTime": {
            "type": "string",
            "description": "End date-time; must be after StartDateTime.",
            "format": "date-time",
            "nullable": true
          },
          "priorityOrder": {
            "type": "integer",
            "description": "Rank relative to other plans (lower = higher priority, ≥ 0). Create default: after the last plan; update default: unchanged.",
            "format": "int32",
            "nullable": true
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanItemRequest"
            },
            "description": "At least one existing task, habit or learning resource; no duplicates. On update the list replaces the items (retained items keep IsDone).",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Body for creating or updating a plan."
      },
      "PlanResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "estimatedDurationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "startDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "endDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "priorityOrder": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/PlanStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "startNotifiedAt": {
            "type": "string",
            "description": "When the start notification was acknowledged (null = not yet).",
            "format": "date-time",
            "nullable": true
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanItemResponse"
            }
          },
          "totalItems": {
            "type": "integer",
            "format": "int32"
          },
          "doneItems": {
            "type": "integer",
            "format": "int32"
          },
          "completionPercentage": {
            "type": "number",
            "description": "Done items / total items × 100 (2 decimals).",
            "format": "double"
          },
          "hasStarted": {
            "type": "boolean"
          },
          "hasEnded": {
            "type": "boolean"
          },
          "isRestTimeActive": {
            "type": "boolean",
            "description": "True only while the plan is in progress (between start and end, items still open)."
          },
          "restTimeSeconds": {
            "type": "integer",
            "description": "Seconds remaining until EndDateTime while rest time is active, else null.",
            "format": "int64",
            "nullable": true
          },
          "secondsUntilStart": {
            "type": "integer",
            "description": "Seconds until StartDateTime for plans not yet started, else null.",
            "format": "int64",
            "nullable": true
          },
          "timeElapsedPercentage": {
            "type": "number",
            "description": "Elapsed share of the start→end window (0–100).",
            "format": "double"
          },
          "serverTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "A plan with server-computed progress. Status, completion and rest time are computed at\r\n`ServerTime`; clients count rest time down locally from `RestTimeSeconds` or `EndDateTime`."
      },
      "PlanStatus": {
        "enum": [
          "NotStarted",
          "InProgress",
          "Completed"
        ],
        "type": "string"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "UpdatePlanItemRequest": {
        "type": "object",
        "properties": {
          "isDone": {
            "type": "boolean"
          }
        },
        "additionalProperties": false,
        "description": "Body for marking a plan item done / not done."
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "additionalProperties": { }
      }
    }
  }
}